// src/types/conversation.ts
export enum ConversationState {
  WAITING_NAME = "WAITING_NAME",
  WAITING_ADDRESS = "WAITING_ADDRESS",
  WAITING_PHONE = "WAITING_PHONE",

  REGISTERED = "REGISTERED",

  WAITING_MENU_SELECTION = "WAITING_MENU_SELECTION",
  WAITING_QUANTITY = "WAITING_QUANTITY",
  WAITING_CONFIRMATION = "WAITING_CONFIRMATION",
}